Refactor `submit` subpackage into `submit.py`
This includes moving the tests into the main `tests/` subdirectory
as `test_submit_remote_client` and `test_submit_cli`.
