+++ distributed/cli/tests/test_dask_remote.py
@@ -1,0 +1,9 @@
+from click.testing import CliRunner
+from distributed.cli.dask_remote import main
+
+
+def test_dask_remote():
+    runner = CliRunner()
+    result = runner.invoke(main, ['--help'])
+    assert '--host TEXT     IP or hostname of this server' in result.output
+    assert result.exit_code == 0
